requirements-base.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Werkzeug==3.1.3
marshmallow==3.26.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-Session==0.8.0
Flask-Migrate==4.1.0
email-validator==2.2.0
requests==2.32.3
boto3==1.36.22



